requirements.txt
pytest==7.4.0
selenium==4.15.0
webdriver-manager==4.0.1
requests==2.31.0
pytest-html==3.2.0
allure-pytest==2.13.2
PyYAML==6.0.1
SQLAlchemy==2.0.21
cryptography==41.0.4
